WIP: Creating source file and errors
